Split Footer and FinePrint components
This splits the Footer component (Description + Filters) from the FinePrint component (Version, Citation, Download).

This allows showOnlyPanels to keep displaying fine print.
